[AUA] Fix the initial state for self upgrade
For a self-upgrade, it is normal for there to be no mandatory steps,
but this leads to the initial state of the tool window to be a state
with no steps selected.  We should default to the assumption that the
user intends to do something, so in that case make the optional /
"Recommended" steps enabled.

Bug: 236260004
Test: added
Change-Id: I3ef5d2fb05352cb08876ab828e268507709a9f44

GitOrigin-RevId: 1603b91e8926164fb3cfc15cea621811380e6d2f