[DeviceSpec] Support more flexible String tokens
Before, the STRING_T and NUMERIC_T would collide and strings that the
user would expect to be interpreted as STRING_T would be partially
parsed to NUMERIC_T and then STRING_T, which causes syntax issues.

This makes it so that the token parsing is done in a more orderly
manner, where we prevent collissions for the relevant parameters.

Fixes: 236302806
Test: DeviceSpecParserTest
Change-Id: I5887cde3be48b179b9d2c70c14153b30b7e0f55d

GitOrigin-RevId: 2739c3533bc492a447a4c7387db5179c080fea0a