Fix leak in LiveLiteralsService HighlightTracker
HighlightTracker keeps a strong reference to the Editor that should show
the highlights. If the Editor is disposed, the HighlightTracker does not
have a function anymore and should be disposed.
This change ensures the HighlightTracker is disposed if the Editor is
disposed or if the LiveLiteralsService becomes inactive.

Also, this ensures that the HighlightTracker does not keep a strong
reference to the editor since the HighlightTracker only can be used with
Editors that are still not disposed.

Test: LiveLiteralsServiceTest
Fixes: 235743082
Change-Id: Ib8fcd5e28889bfde2037651cf9e07bc3670fb5fa

GitOrigin-RevId: c23ab2eceeb546d12bbc396ad7f40f1b70977ee3