Do not switch to UPP automatically when switching between files.
If a design file is opened already, switching the focus from another
file to that design file should not switch the tab of problems panel
to universal problems panel (a.k.a. UPP).

This only applies to the opened file. If user tries to open a file
which is not opened yet, the selected tab of problems panel will
automatically switch to UPP.

Bug: 236385632
Test: update and covered by IssuePanelServiceTest.
Test: manually.
Change-Id: I4c062313b3d3fc86928084be2d554bad1a0d3900

GitOrigin-RevId: a990d1a70446ff1c516f0f40f88b85c11c5a963b