Fix Application service for AdbLibSession
Ensure ADB is started from the default SDK location in the
`AdbLibSession` returned from the `Application` service.

This ensures that the AdbLibSession has access to ADB even if
there are no projects open, or when the ADB server is stopped
(manually or from inside Android Studio).

This behavior is required for the ddmlib compatibility layer, as this
layer requires the Application level AdbLibSession to work, since
ddmlib only has global state and does not know anything about
`Project` instances.

Bug: 188067646
Test: n/a
Change-Id: I2461b228238d0799d2da5dd989c30b0db5a556d1

GitOrigin-RevId: 3f103268196d757e145d41f79609d651e32b3be5