Remove LintClient from IdeGooglePlaySdkIndex
The lifecycles of these classes don't seem to match,
IdeGooglePlaySdkIndex is a singleton whereas LintClient is tied to the
lifetime of the Project.

It doesn't seem to be required in this class anyway, the function where
it is used doesn't seem to be called currently and I'm doubtful these
errors need to be passed to the LintClient either way.

Bug: 237114637
Test: N/A
Change-Id: I5b6166ca2855084458cf52f8fa49df6c64b19899

GitOrigin-RevId: fff044562216432ff0355f9c1cfe081da0a9e909