[Dsl,.toml] Handle conversions between version and version.ref
Define a local subclass of GradleDslLiteral for version and
version.ref properties, with overriding methods to handle the special
cases of converting between the two kinds of version specifications,
and to convert new reference values to the appropriate literal syntax.
This allows removing the abstraction violation from the TomlDslParser
convertToPsiElement, as nothing should be a ReferenceTo by that stage.

Fixes: 237659919
Test: added
Change-Id: Ie968c2d20fc667594ada1b89bcc929999bf6073d

GitOrigin-RevId: 032d01ff6472ec2e3035743dffa5f4668125d271