Fix bug with regex matchers
Whenever a SwingClassRegexMatch was specified, the component it found
was returned along with its entire subtree. This made it impossible to
identify a single component using this kind of match when that component
had children.

The fix is to return the component by itself when the regex matcher is
the final matcher specified.

Bug: N/A
Test: this is for the test framework
Change-Id: Ic50ea0ab12e049f6bb30820c7c266cc68135b097

GitOrigin-RevId: b83c1233e7daaea77cba64fb8e2f1ca10346973e