Do not use runBlocking in updating previews
runBlocking should not be used within the suspendable method that
updates the preview. This can lead do dead locks if the refresh runs on
the UI thread since it executed block might wait on the UI thread to be
freed while it is waiting for runBlocking.
This changes the code to use withContext.

Test: Manually checked
Fixes: 238735830
Change-Id: Idcd5a19e2b083756d622f06a7d15822f9686e783

GitOrigin-RevId: ef9aa3250c6fcac054b353a55b2fdcd4fcae834a