Allow AndroidSystem to be a rule
If a test times out, Bazel would issue a SIG_TERM and Java will not
execute the finally clause, thereby not closing the video recording
properly. When video recording doesn't close properly, the resultant
video would be unviewable.

This commit turns AndroidSystem into an optional Rule, with default
TempFolder creation if none is given. This allows the video capture
to shut down properly outside the body of the test (in @After).

Bug: N/A
Test: generated video files in a timed out test are viewable
Change-Id: I784e77aa33fdbb561db5c9c9df293a85f667a6cf

GitOrigin-RevId: 6b2668ccf9afa47ca20d9ec681bce14a27da7e27