Create new device mirroring benchmarking app.
This app allows us to time a round-trip from input in
Studio to the device and back. When a touch input happens
on the screen, those coordinates are encoded as two
colors in the display which can be read in studio.

To ensure that the video encoder is also exercised
appropriately, some random noise is generated and a visual
element moves around part of the screen relative to the
touch input.

Test: Nope
Bug: 238805426

Change-Id: Iac0b495329ffc7fd701f3fb3cd8d481efbec46dc

GitOrigin-RevId: bed6a67e33adbfe296ddb43d6994a0a0036b1eb1