[Dsl,.toml] Set the writing support flag to true
Now that the writer backend supports what the PSD uses, and we have UI
in the Experimental Settings dialog for users to turn it off, we
can (relatively) safely turn on write support by default in the
expectation that what gets written is a reasonable attempt at what is
expressed.  There are still missing features at higher
levels (e.g. the Ps layer doesn't have a concept of Version Catalog
version variables, which would be good to have so that the PSD can
provide UI for them) but what is supported should be reasonably sound.

Bug: 200280395
Test: existing
Change-Id: Ibb7cae5bb5669a9278643c46073d693aa872f4e0

GitOrigin-RevId: 3df571fc9c5fe43fb0886c52e70b4f01920f70ca