Remove assertion of deleted files
deleteRecursivelyIfExists is sometimes not able to delete
files that are perhaps in use and throws a
DirectoryNotEmptyException. The exception causes false
negatives in the test as the test is not able to clean up
properly despite the test body succeeding.

This commit works around this problem by manually crawling
the directory structure and not asserting if a deletion
attempt has succeeded.

Bug: 239343337
Test: no more cleanup issues in 500 runs
Change-Id: I1f8c7e62a4e72e410e1b90ccc9474bad2f1189ee

GitOrigin-RevId: eb0c2832d380aa45617f1bc1c7365c71ab2744c7