Remove `ProfilersApplicationRule`
We use to have a custom `ProfilersApplicationRule` as a test-time hack
for running everything in the main thread. Following change Ic0656ffb,
we remove this hack, and explicitly pass a "runner" to each component
that can run things in the background, so that in tests, we can run
the work on the same thread. Compared to the "waiting for async
updates" solution, this eliminates flaky tests.

The root that needs to be parameterized is `CpuTreeModel`, and the
process propagates through the following classes until we reach a
class that has access to `IdeProfilerServices.poolExecutor`.
  CpuTreeModel
  -> CaptureDetails.{Aggregate, TopDown, BottomUp}
     -> CaptureDetails.Type
        -> CpuAnalysisChartModel
           -> CpuThreadTrackModel
              -> CpuCaptureStage ✓
           -> CaptureNodeAnalysisModel
              -> CpuThreadTrackRenderer
                 -> ProfilerTrackRendererFactory ✓
              -> CpuFullTraceAnalysisModel
                 -> CpuCaptureStage ✓
              -> JankAnalysisModel
                 -> JankyFrameTrackRenderer
                    -> ProfilerTrackRendererFactory ✓

Fixes: 240757470
Fixes: 240756948
Fixes: 240716875
Test: update existing; run 100 each locally
Change-Id: I643e0a18e289f3e0e0d02433229daac5281f67e5

GitOrigin-RevId: d4ba0fb28aa107711d1deb21375496edf24c36ae