Fix do not display custom view tab in shared issue panel
Before we treat all kotlin file as Compose file. If user writing a
custom view by using Kotlin file, the name tab is wrong.

Moreover, shared issue panel searches the DesignSurface when file
editor is changed. However, the design surface may not be initialized
yet when editor is just changed.

In this change, the designSurface is registered with its file into the
IssuePanelService. This avoid the potential race condition and have a
chance to recognize the given surface is for compose or custom view.

Fixes: 238810669
Fixes: 238789488
Test: covered and add regression test
Change-Id: I032cc6294263d4abd8ee8c422befcd1e10590951

GitOrigin-RevId: f6c81dc4878a87815cc9287a5ecebe2ab0f1e873