Fix sym-link handling at project setup
When a project is referred to via a symbolic link Gradle returns its
root as a resolved path. This may result in `rootProjectPath` property
in modules in the root build not matching the `buildDir` of the root
build in Gradle.

We used to correct `rootProjectPath` when using it to initialize
`buildId` values when requesting sync but we forgot to do the same when
translating task names from included build relative task names to root
build relative task names.

(1) Revert the fixup applied to `rootProjectPath` at the usage side.

(2) Configure Gradle project root as an already resolved path so that
    it matches the path returned by Gradle.

(3) Add more tests for the following cases (a) project (build) root is
    a symlink; (b) a non-root Gradle project is included via a sym-link;
    (*) a case with a build included into a composite build via
    a sym-link is not currently supported by the Android Gradle plugin.

Bug: 129044665
Test: n/a
Change-Id: I8a26ea332ee81788abc65e5963ec5b51250cf5cc

GitOrigin-RevId: e2c949848c30cb41363794bb3f6c64f7f640925e