[BA] Fix task detail page creation scroll bug
For a long time we had this bug that some detail pages when opened
for the first time are created scrolled to the bottom of it.
Currently, it looks like only the tasks and plugins details pages
are affected. In this change I talk only about tasks, plugins fix
will follow up in the next CL.
Task details is broken only on the tasks view, not on the Warnings
view even though the implementation is the same for both places.
I carefully checked that the only significant difference that
triggers this behaviour is the presence of the tasks chart on the
tasks view, manually verified that it works properly without it.

The precise reason for this remains unclear, from the debugger I see
that in faulty case page size is set before the caret is updated in
html panels making panel to scroll down to make last element's
caret visible. On the Warnings page same caret magic happens before
setting any sizes and does not affect the scroll position.

In order to mitigate this effect do the following:
 1) change task page to be a single html panel instead of set of
 multiple components. Additionally it simplifies layout and makes it
 easier for the user to select and copy the whole thing at once.
 2) force reset caret to the beginning of the element after inserting
 text so that it does not scroll down to make it visible.

Additionally reimplement test for the task details panel to make it
assess the full page generated html for all available options.

Bug: 161898008
Test: updated TaskDetailsPageTest.kt
Change-Id: I318ac8dbdb23987d899d4b7ca8e33c66c5a0e41b

GitOrigin-RevId: 8eefe22f04c682d2666a0d764c795ba06f17bfea