[SDK Index] Use StudioFlags in PSD initialize
The SDK index can be initialized when used by lint or the PSD. After it
is initialized by lint, the StudioFlags are used to configure it but
that part of the initialization is not being done when PSD uses it,
causing different behaviour when lint is used and not.

This CL causes the flags to also be used when the index is initialized
in PSD by now using IdeGooglePlaySdkIndex#initializeAndSetFlags in both
places.

Bug: 239749114
Test: Covered by test spec of SDK Index.
Change-Id: I5838d137f83aacf94e84ac6ca0f41aa7e989fe1e

GitOrigin-RevId: 0ca251d4172dd9124c4a517eca6ef4abe7b06de0