Fix intermittent toolbar in Compose Preview
We have a toolbar (ActionsToolbar#centerToolbar) that is used by layout
XML files to display some actions depending on the component selection
in the layout editor.

For Compose Previews, this toolbar was supposed to be empty. However, a
loading icon is shown before the toolbar detects it's supposed to have
no actions, resulting in a glitchy UI (the toolbar shows up and shortly
disappears).

This change fixes this behavior by only adding the centerToolbar for XML
files.

Bug: 238291099
Test: Visually verified. ActionsToolbarTest should still pass.
Change-Id: I01f7e319fd453f436950c6d06466f0e27d36102e

GitOrigin-RevId: 62a2b4862b524b7268bb4d9cb678e850c5e4b475