Accept JavaFileType in AndroidPositionManager
AndroidPositionManager's getAcceptedFileTypes() method is called by the CompoundPositionManager that contains it, in order to determine whether it should be used for any given file type. However, it appears that in the past this value wasn't used correctly, and so AndroidPositionManager has been advertising support for the wrong type.

CompoundPositionManager used to incorrectly determine the file type of a source position being debugged; as a result, it would go through all of its child position managers regardless of what file types they supported. This was fixed upstream a few months ago with https://github.com/JetBrains/intellij-community/commit/590ab29b60891141a02831094ff1f2df034fca68.

With the above fix in place, it's now apparent that AndroidPositionManager is supporting the wrong file type (and likely always has been, since it hasn't mattered in the past). Both when debugging into SDK sources and setting breakpoints in those sources, the locations resolve to a "Java" file type rather than a "JavaClass" file type.

I can't say with 100% certainty that there's never a scenario where the class type could be invoked, but I haven't been able to find any. (This has included debugging through multiple various scenarios with breakpoints enabled to trigger if/when a JavaClass was sent to the position manager; none were ever hit.) I specifically tried out debugging into the SDK when no sources were available so that a decompiled version of the SDK coming from compiled class files would be used; but even in that scenario, the locations in question are represented with a "Java" file type. Given this, it seems likeliest that the JavaClass type has simply always been incorrect, so I've opted to replace it with the Java type instead of supporting both.

Test: Covered by new unit test AndroidPositionManagerTest#testGetAcceptedFileTypes_acceptsJavaFiles
Bug: 74128300
Change-Id: I48e91d5f41135befa26a1e0241ded6128c528fff

GitOrigin-RevId: bb5086fde93251547fb9c4d3cb55d6b18cc7d1f1