Report data service crashes as failed sync back to caller
Exceptions in the data services phase of sync are reported as failures
in `GradleSyncStateImpl.kt`. However, when sync is requested by the
Android IDE plugin itself, results are reported back via
`IdeaSyncPopulateProjectTask`. It allows to bind requests to results.

The later execution path does not handle exceptions in the data services
phase of sync yet.

Bug: 239800707
Test: PlatformIntegrationTest
Change-Id: I5999d8ff8f98e1c6ea62ce9f4e495622a5be0925

GitOrigin-RevId: 24a525932e314bd48af49384fd6b9e3a2e8aad45