Make sure the recomposition visibility is set
The function updateRecompositionColumnVisibility is called each
time a model update is received. The function will be hiding the
recomposition columns if we are not connected or the setting is
off.

Unfortunately the 1st model update may arrive before the client is
marked connected thus hiding the recompotition columns when they
should be visible.

Start listening to connection updates and perform another
updateRecompositionColumnVisibility on each change.

Also: The work inside updateRecompositionColumnVisibility should
be executed on the UI thread to avoid potential drawing problems.
(the link was sometimes drawn twice).

Fixes: 240122281
Test: Added unit test
Change-Id: I3c192280ad16ff8a3fdb26903d96237606af9ff6

GitOrigin-RevId: df6a02fcfcf3b9d39942fd7ef78e5fd6df688378