Publish sync results to project system topic reliably
The existing implementation, which publishes project system wide events
from a `GradleSyncListener`, relies on subscribing to
`GRADLE_SYNC_TOPIC` events at the right moment during the initialization
and it is difficult to maintain.

Thus move publishing to `PROJECT_SYSTEM_SYNC_TOPIC` directly to
`GradleSyncStateImpl` and publish to this topic after publishing to our
internal topic.

Bug: 133154939
Test: PlatformIntegrationTest
Test: existing tests pass
Change-Id: I768d87d25493ff167a5fcada5f79f10bba6afb6d

GitOrigin-RevId: 23883a7de7712aa126f6a5178d015f9fdb9675ba