Fix race condition in logcat panel
There was a race condition in the logcat where we report having a device
before we have started the service. While this did not present any
problems in production, the race condition allowed for failures in our
tests. See ag/19419228 failures for an example of an unrelated change
seeing test failures.

The failures could be reproduced by running connectDevice_readLogcat on
its own (fail) and running it as part of the class (pass).

Change-Id: I98844228a3ce94a3a082c33db44c18c10c9620b5

GitOrigin-RevId: 5fb7ca1f13b2b52e377e6b6ff96ac1d6c2628869