Handle KotlinGradleModel and KotlinMPPGradleModel
both present when preserving user data passed by the Kotlin IDE plugin
outside of the data node tree. The workaround is a temporary solution
until the problem is fixed in the Kotlin IDE plugin. However, the
workaround makes the wrong assumption that if both models are present
their `cacheOriginIdentifier`s are the same.

Preserve data for both identifiers instead of asserting they are equal.

Bug: 240400766
Bug: KTIJ-22303
Test: existing
Test: testKotlinMultiplatform_hierarchical_withJs
Change-Id: If643e1b6651c45b42666869e0c09d915008b12de

GitOrigin-RevId: 65dce82df05bef0d6e23d4680a601a065763bbfe