Pass Delay as a Parameter to LogcatServiceImpl
Logcat format is not well-defined. A log entry is terminated by a "\n\n" but it can also contain a "\n\n" inside a single log entry. So there is no way to tell when an entry ends. This means that we cannot consume the last log entry immediately, we need to allow for the possibility that another batch is arriving from the socket. If another batch doesn't arrive within a certain time period, we consume the entry.

By definition, this is flaky but there is no other alternative that I am aware of. I have plans to switch to the binary format which is more robust, but it's also more fragile, so I'm not sure when that will happen.

The test is failing because LogcatMessageAssembler is consuming the last message from the server too early, before the server terminates.

By passing a large delay value to the assembler, we can prevent that from happening.

http://sponge/ed5c43dc-1862-4996-bd2e-377b590fa4be

Fixes: 240555493
Test: Self
Change-Id: Id4be6052aee6a9f137d3279ef34623cfa414b6b2

GitOrigin-RevId: 4d3fc0ef69b0250637e04b09c80422290e8797c2