Reuse the preview editor for issues of same file in shared issue panel
The linters analysis the new issues when file is changed by preview
editor. This makes the issue updated, however, if the selected issue
is gone, the detail panel and its preview editor is removed. This makes
user cannot keep focus on the same preview editor.

This change avoids creating duplicated preview editor in shared issue
panel. Modifying the preview editor makes the tree be updated and grab
the focus. The following changes should keep the focus on the editor.

Bug: 236393984
Test: manually and update the test cases.
Change-Id: I6b4f378a048d4fe0827adb40d34885862907ce31

GitOrigin-RevId: d923733753f265362549153b03a43736c800963c