Introduce GradleSyncListenerWithRoots
In the IntelliJ platform multiple Gradle projects can be attached to an
IDE project at different locations and none of them is required to match
the root of the IDE project.

Android Studio used to make an assumption that these roots are the same.

Begin introducing explicit `rootProjectDirectory` in listeners and
state management as a step towards correct support of arbitrary Gradle
roots.

(1) Introduce `GradleSyncListenerWithRoots` and use it instead of
    `GradleSyncListener` in `GRADLE_SYNC_TOPIC`;

(2) Keep using `GradleSyncListener` as means of communicating back
    the status of a requested sync since the root is known to the
    caller anyway;

(3) For now, keep using for `GradleSyncListener` for external
    subscriptions;

(4) Fix invoking Gradle sync on projects with a Gradle root attached
    to a project sub-directory.

Bug: n/a
Test: PlatformIntegrationTest
Change-Id: Ife2e65fa717d60acbfdb4b0c39a30ba7b68f9b46

GitOrigin-RevId: cb37e71ea91bbad2a688ad783d43f7d22c766981