Do not refresh VFS to find build files
Build files are located under content roots and therefore should be be
valid virtual files after syncing.

If the project tree is refreshed when any of these files are invalid we
will skip these files. They should re-appear when the VFS is refreshed
since the project tree view is automatically refreshed in this case.

Bug: 206829505
Test: n/a
Change-Id: Ic5145fbb878b96ce6ff5ce57ba86815569ed9c5b

GitOrigin-RevId: b8fbd93226a8ec48f755d9d287e721e524a07dfe