Prove build output dirs are not refreshed
Output directories containing `.class` files usually are excluded from
the project structure. Nevertheless, they are often used to load/inspect
produced `.class` files and are expected to be present in the VFS when
Gradle build completes. Android Studio, unlike IntelliJ IDEA platform,
does not ensure that these directories are available in the VFS on
build completion.

Add a test illustrating the problem before it is fixed.

Bug: 74386986
Bug: 241686649
Test: n/a
Change-Id: I413f4fe268f816e7d7d460b8ca0f1a37863e132c

GitOrigin-RevId: 989bc55b64ed1ab6a8b4ba101273b68db71d0208