Fix flake in SessionStatisticsTest
The stats.reset() was executed after the model was notified to
update. If the background thread finished updating the measurements
before the stats.reset(), the measurementCount would be reset and
the test would time out.

In addition the update notification was sent out before the model
updating flag was reset, which gives a small chance that the
memory measurements would be cancelled (because the logic would
cancel the measurement if another model update was happening.)

Fixes: 242197596
Test: SessionStatisticsTest
Change-Id: Ifdf38e0f6cbeb7674f0a8bc3db9f63ee4a4fcc4c

GitOrigin-RevId: 495b799f695a44e44cdcade09d6619a9c6635e37