Create new AndroidPositionManager behind flag
The intent of this change is to allow both an "old" and a "new" version of this file to exist in parallel as feature development is done on the new version. The old version (AndroidPositionManagerOriginal.java) is not expected to have any new changes, and will eventually be deleted once the new version is ready.

A previous CL did the actual copy of the file; this change fixes up the "original" file with an appropriate class name and comments, and adds the flag to decide which version of the manager should be used.

Change-Id: I522b34a5d30ee14732538e40dcf27d1293544cfc
Test: Included
Bug: 74128300

GitOrigin-RevId: 4d36f935e8af24b7c2c5c8e01925126fc17ea2c9