Introduce AndroidGradleProjectConfigurator
which implement `DirectoryProjectConfigurator` and configures the root
module using `TopLevelModuleFactory.createOrConfigureTopLevelModule()`.

When opening a project without modules the IntelliJ platform uses
`DirectoryProjectConfigurator` to perform the initial setup. The default
configurator simply creates the top level module.

Android Studio similarly creates the top level module when opening
a project as additional settings need to be applied. However, this
happens only when opening a project without an `.idea` directory and not
when a newly checked out project is opened, which usually would have a
`.idea` directory but no modules.

In upcoming changes we will (1) configure top level modules for all
Gradle roots found in the project and (2) stop doing this directly in
the open project processor in favor of
`AndroidGradleProjectConfigurator` only.

Note: the root module name in `AS36_noiml` is now determined by the
      `PlatformProjectConfigurator`, which uses a different location
      for storing `.iml` files than the `GradleManager` does.

Bug: 200820556
Test: n/a
Change-Id: Ia4aeeb4fa4bafda1c2dd7164fcd14bc660526ece

GitOrigin-RevId: a2c0a0c3d6ef7a96da2833bd9dc4ccfbc5ba2f8b