Configure top module per Gradle root
When opening a project without modules `DirectoryProjectConfigurators`
are used to set up some initial user experience.

Android Studio creates a top level module and links it to the Gradle
external project to avoid later re-configuration, which might result in
unnecessary long unstable module names. Having a top level module allows
users to browse the project while it is still being synced with Gradle.

In IDEA and as a result in Android Studio it is possible to have an IDE
project to which multiple Gradle roots are attached.

Create a top level module for each Gradle root so that configuration
files from each root are visible in the project tree view before sync
completes.

Bug: 242440055
Test: *.testSimpleApplicationMultipleRoots
Change-Id: Id0ff3238a8f88c14241bd5e8fd91fdaee766e5df

GitOrigin-RevId: dcf17cfb84f79a9e3fdb6f615c6aa8257930e6d2