Prevent Disposable leak on AndroidCoroutineScope
Since AndroidCoroutineScope constructor function allows passing
an optional context that can override the `Job` of the newly created
CoroutineScope, it is possible to leak the parent `Disposable` attached
to the default `SupervisorJob` instance.

For example, calling `AndroidCoroutineScope(disposable, MyJob())`
results in `MyJob` not being automatically cancelled when `disposable`
is disposed. Instead the job of the default `SupervisorJob` will be
cancelled, which is not useful since the default `SupervisorJob` ends
up being leaked and not used in this example.

The fix is to cancel the `Job` of the created coroutine, ensuring
we attach to the `Job` of the optional context if needed.

Bug: n/a
Test: n/a
Change-Id: I21593f7250ebcb270eabd2a88526a457946b2d97

GitOrigin-RevId: ee494938e9cd4572899b67bc8a53a23552f11714