Use per-device setting when auto-connect is on
When auto-connect is on, we don't want to set the per-app setting "debug_view_attributes_application_package", as this would cause every activity to restart each time the user navigates between app.

Instead we want to use the per-device setting "debug_view_attributes".

This change uses that setting when the flag for auto-connect is turned on.

The per-device flag should be turned off when:

1. The device is disconnected.
2. Layout Inspector stops.

This change makes `DebugViewAttributes` a singleton, as the per-device flag has to be set and cleared in different locations, but using the same instance of `DebugViewAttributes`.

In the current state the per-device flag is turned on when:

1. Auto-connect is turned on and the use enables the run configuration.
2. Layout Inspector connects to a process.

The per-app flag behaviour remains the same.

Test: LayoutInspectorLaunchTaskContributorTest, ForegroundProcessDetectionTest, AppInspectionInspectorClientTest
Bug: 241970079
Change-Id: I331a97e84f10cfd7ac45f834237688988b332e65

GitOrigin-RevId: 0a227440e05a3cbbbbd00d169d42912b38491a95