[AUA] Match AGP's option parsing for properties
For character sequences (as come from properties files), AGP ends up
comparing the lowercase (in US locale) of the value against the string
"true", so replicate that there.  The other code paths (non
CharSequence branches) in OptionParsers#parseBoolean are used when
"parsing" from project extra properties (not applicable to this
refactoring processor).

Bug: 189366810
Test: added
Change-Id: I1d74157965ed2ba32bc02bc7f745ce67040ec508

GitOrigin-RevId: dfca41a93d3f79f32d6cdfa09686f28101e6c7cd