Fix support for manual test fixture artifacts
in Android modules.

Dependencies on Android project custom configurations specified as

```
    implementation project(path: ":lib", configuration: "manualTestFixtures")
```

are reported by the AGP as module dependencies with a not null value of
the `artifact` property. However, the AGP may also report dependencies
on Android components in the same way in some edge cases related to
tests. It makes it impossible to decide whether a dependency targets the
`main` artifact of a Gradle project or a custom configuration producing
a `.jar` file.

Assume that if a resolved dependency gets a build type or a product
flavour it is a dependency on an Android component and ignore the value
of `artifact` property even when it is not null.

Bug: 218883113
Test: testNonStandardSourceSetDependencies_manualTestFixturesWorkaround
Change-Id: Id23fa67e77735893957192908f7def06f286f8c8

GitOrigin-RevId: 839b59973772428e793326030a6643b9f90c1f5d