Initialize low mem notifiers using PostStartupActivities
AndroidLowMemoryNotifier and LowMemoryReporter are currently
registered as application components. At application startup,
they register themselves to listen for low memory notifications.

This CL moves both of these classes to be application level
services. AndroidLowMemoryNotifier is initialized from a
post startup activity (=> it'll be initialized ~5s after a
project is open). We want to capture metrics from
LowMemoryReporter, so it is initialized a little earlier
when the application starts.

Bug: 243082216
Test: manually tested that
   a) the services are initialized only once
   b) they are initialized only after a project is opened
   c) they are disposed when the application is disposed

Change-Id: I856c3a1d946c802f8e173b1918d6dfb3d7984552

GitOrigin-RevId: 79bbea55b8f084a3e94fd554839dbe132e922af6