Remove tweakDefaultColorScheme()
This code was added in http://aosp/97902 and was solving an old
problem in which XML tags were highlighted with an off-white
background color (which did not look very good).

However, as described in http://b/145750280#comment6, IntelliJ has a
new default light theme. This light theme uses the "IntelliJ Light"
editor color scheme by default, which no longer suffers from the
original issue.

Users will still see the off-white XML tag background color if they
manually select the "Classic Light" color scheme---but since this is
not the default, it is not worth modifying it just for Android Studio.

Bug: 145750280
Bug: 200037669
Test: verified that a fresh install of AS uses the "IntelliJ Light"
  editor color scheme by default (when using the light IDE theme) and
  that this color scheme has normal XML background colors.
Change-Id: I2ebaa25bead98fa3a5eabd2b3d52c01eac958342

GitOrigin-RevId: 46d729bfae4f64c52c4d9d5829902271657ec91f