[SDK Index] Add E2E for PSD integration
... similar to lint SdkIndexLintTest.

SDK Index initialization can happen when a build.gradle file is open in
the editor and when PSD is shown. We currently have a test to confirm
the initialization is done by lint when opening a build.gradle in the
editor. This change adds a similar tests but instead of opening the
editor it opens the PSD.

Bug: 233789469
Test: SdkIndexPsdTest, SdkIndexLintTest
Change-Id: I392afca72aab2ebf21812a12baed181955e4f2b4

GitOrigin-RevId: f6d5d9df0b418d035e802fd61de99a22f3da6491