Init GcPauseWatcher independent of AndroidStudioInitializer
This CL converts GcPauseWatcher to be an application service,
and triggers its initialization from a background startup
activity (which runs a few seconds *after* a project is
opened). The primary goal however is just to reduce what we
do in AndroidStudioInitializer so that we can eventually remove
that application component.

Bug: 243082216
Test: open project in Studio and verify GcPauseWatcher is
  initialized
Change-Id: I680d3215c3363b6402c3af18b5f2833fd621a39a

GitOrigin-RevId: c0dc1a9e07f484f855b18ae0ecab9d108e57e112