Move ServerFlagDownloader initialization out of AS Initializer
ServerFlagDownloader is responsible for triggering a download
of the latest flags from a server. This is not time sensitive,
and it doesn't have to happen right when the app is launched.

This CL moves this operation out of AndroidStudioInitializer
into a post startup activity in ServerFlagDownloader.

This CL also includes a couple of minor changes:
 * ServerFlagDownloader is triggered just once, so there is no
   need for a scheduled executor and we can just launch it
   immediately on a pooled thread.
 * AndroidStudioSystemHealthMonitor used to depend on
   serverflags until a prior CL (ag/15806791) made ServerFlags
   an application service. This CL removes a stale comment.

Bug: 243082216
Test: launch Studio, verify ServerFlagDownloader is triggered
 after a project is opened.
Change-Id: Ic0610f63aec0a5ba00a9ec853fcf4ebc8020dd8f

GitOrigin-RevId: 44f0b5aafe76bd57ea6e160e9959399dde300eef