Renaming myShowUnitAtMax variable and its setter
The boolean variable indicating whether the profiler
chart y-axis labels should only show the unit for the max
value was named "myShowUnitAtMax". We believe a more accurate
name for the variable would be "myOnlyShowUnitAtMax".

This variable was thus renamed to the proposed name above
and so was its setter method from "setShowUnitAtMax"
to "setOnlyShowUnitAtMax". Also added javadoc comment
to the instance field declaration to clarify meaning.

Bug: N/A
Test: N/A
Change-Id: I546c8216b3b2349764ae6312fecfd561f46c8032

GitOrigin-RevId: 756d8fa1625ebb11cca23b270ed03bdb5b9138a6