Simplify AdbFileProvider & remove early initialization
The logic to find the location of adb relies on AndroidSdkUtils,
and hence it resides in android-core module. In order to keep
android-adb independent of android-core, we currently use a
scheme where application startup and project listeners explicitly
stoer an adb provider into the userdata of an application and
the project respectively.

This CL simplifies this mechanism by utilizing the fact that
we can register application & project services with an interface
defined in android-adb, but implementation provided by core.
This cuts down on the early initialization and the reliance
on userdata fields.

Bug: 200037669
Test: existing tests
Change-Id: I5c4733fa51eaaed7a6023e242e7591448e84c2e0

GitOrigin-RevId: b60de92d9d6ab7bf6d7e59cbf56d09f0accab603