Use \n instead of <br> in sync messages
The build output window does not recognize <br> when it appears in the
first line of a message and shows in the tree view as it is. Other parts
of the pipeline treat the new line symbols and <br> in the same way.

This change makes it possible to unify processing of sync messages and
avoid two different ways of reporting, which in turn makes it possible
to attribute sync issues to the correct Gradle root.

This change is unfortunately manually tested only. Upcoming changes
aimed at better testability of this code include a test that ensures
this behavior.

Bug: 242440055
Test: n/a (manually verified the output is the same when this code runs)
Change-Id: Id65e42307989b5146a3795eb1a8c2b80a0df5590

GitOrigin-RevId: 785de374a2767ff80a423d59cd27b3dec7348843