[AUA] Update integration test for 8.0.0-dev
Now the manifest package to dsl namespace processor becomes mandatory.

This actually reveals a dependence in our tests on the order of
processor declarations: because we implicitly test two-stage upgrades,
we must have that the order of applying operations from initial to
final state directly must be the same order as initial to intermediate
then intermediate to final.

In practice what this means is that the components in the
AgpUpgradeRefactoringProcessor should be sorted by when they first
become mandatory, with a tiebreaker of when they first become
optional.  (For now just move some 8.0 processors before the 9.0 ones,
but we will have to look at enforcing this.)  Additionally, processors
should avoid creating models in findUsages() that might later be
populated in performRefactoring(), since all of findUsages() executes
before any of performRefactoring() and so the presence of the
yet-to-be blocks can make the ordering of results unexpected.

Also add a minimal `gradle.properties` file to some Tool Window tests
in order to handle the fact that the R8 Full Mode refactoring will
fire on an empty gradle.properties file.  Adjust the expected output
of some other tests to compensate.

Bug: 200682321
Test: existing tests updated
Change-Id: I727bd77a5b4d29f1a16c862e9c6388392d0d11dc

GitOrigin-RevId: 3e8a6ac83c18c527384c32344416c9d1f1b6f201