Add a test for Unit tests with coverage
This change add a test that, given a file from the project, checks if
the JavaCoverageEngine will require project re-compilation.

This is because the way that JavaCoverageEngine works is that, for files
that are open in the AS editor, it verifies if the project is up-to-date
or not by checking if the VFS is up-to-date or not.

Bug: 241426924
Test: added
Change-Id: Idbac6b239a8b2541ba4a2e01359621631717904a

GitOrigin-RevId: 600bb31967d9493b668bf82e48905eab65a7d75b