Add a test for Unit tests with coverage
This change add a test that, given a file from the project, checks if
the JavaCoverageEngine will require project re-compilation.

This is because the way that JavaCoverageEngine works is that, for files
that are open in the AS editor, it verifies if the project is up-to-date
or not by checking if the VFS is up-to-date or not.
(cherry picked from commit c68027188a71d0616567fcd381c15b335003bfe7)

Bug: 241426924
Test: added
Change-Id: Idbac6b239a8b2541ba4a2e01359621631717904a

GitOrigin-RevId: d22f28d5779af8da6870fb20633d4b41d2ed6be4