Disable BLeak DisposerCheck
DisposerExplorer throws NPEs due to commit
d0661356b9778aa21a5f2593b16ecbae8859ef3f, which changed the internal
structure of the Disposer. However, this may get reverted due to
performance impact. This check is not particularly important, so it's
easier to temporarily disable it than change DisposerExplorer's
implementation just for this.

Bug: 243857696
Change-Id: I555374ca98e9099bc1b25ed651a9a8bc7ce96e5e

GitOrigin-RevId: 66793e74f1c82d30d7b660ced5e8322f5cdebd92