Simplify version logic in AndroidPositionManager
AndroidPositionManager is currently written in such a way that it can
support multiple Android versions at the same time. However, a new
instance is created for every new debug process, and each debug process
only has a single device on a single version.

This change simplifies the logic a bit by fetching the version only
once, and looking for a source folder only for the version being used.

Change-Id: I4af47af2c985878f4ae740bacc78e169f8689912
Test: Included
Bug: 74128300

GitOrigin-RevId: 4531b5e907fdad0ed45c8237004b82fd1da92235