Avoid Layoutlib automatic re-renders for Compose Preview
The Compose Preview design surface uses the LayoutlibSceneManager
underneath, which triggers some automatic re-renderings, for
example when a model or its configuration is updated in the
DesignSurface. However, Compose Preview manages its own render and
refresh logic, and then the automatic renderings triggered by
LayoutlibSceneManager are not needed, and moreover, they are
causing some serious problmes. For example, if a user has a file
with N Previews that have the same render configuration, then
a full refresh of them, cuased for example by fast Preview, is
silentely triggering N^2 of the mentioned automatic re-renderings.

This change modifies the Compose Preview design surface to use a
LayoutlibSceneManager instance that has its automatic re-rendering
logic disabled.

Fixes: 243970018
Bug: 240202447
Bug: 243578720
Test: manually verified.
Change-Id: If82709b3040be24cd9f8ef8327b31607152f29c1

GitOrigin-RevId: 35a5f613a1ac281cffc029a797e2b1f1a031ed74