Dispose WindowRecomposer.animationScale in Compose disposer
This map will hold a reference to the BridgeContext causing a full leak
of the ClassLoader and potentially all the View hierarchy as well.
This manually calls via reflection the clear of the field to ensure is
cleared properly.

Bug: 244234828
Test: SingleComposePreviewElementRendererTest
Change-Id: I6f704caf65a2b69a1314d11346fae13f982fb5c7
(cherry picked from commit 463e8549c9c7efaa2d189b4f4a2488995054cb0f)

GitOrigin-RevId: e4cf48bebbc4413860d36e4a27d2956932f5aaed