Reduce the number of times KMP models are built
When fetching Kotlin and Kapt models, we need information
if this is a KMP module. This causes KMP models to be built
and it is not a cheap check. This CL computes this value
only once in order to optimize performance.

Bug: n/a
Test: existing
Change-Id: I76a052e1663ea9346705b6829062620352b74d43
(cherry picked from commit 36b7bba153c5b7684e62350a874a9382dcc8f71e)

GitOrigin-RevId: e86c895656d9451c1e4f6ddd18809d70b1cddc2d