Use concurrent map to store the issue and source in shared issue panel
Creating immutable map uses the iterator, which may still cause the
ConcurrentModificationException. Force using the concurrnet map
instead of the map.

The source to issues map is not modified frequently, using concurrent
map should not hang the ui.

Also compare the issues as the sets in the unit test, since the order
of issues is not kept.

Fixes: 240487574
Test: covered
Change-Id: Ie2b2c7ffc8e77b9e332481d6a66fd046f52cfb35

GitOrigin-RevId: 26fc254f6e66e7aa752dd87ecdcda532796b1125