Fix accessibility of the Wear Pairing panel
The panel will contain a list  of rows but unfortunately, the screen
reader would not be able the text of each one since they are wrapped
in a JPanel.
This change makes sure the panel is correctly tagged to be access
by the screen reader.

Test: Manually verified using MacOS voice over
Fixes: 220874605
Change-Id: If7cee94369a9932ef488ae748f185383cbe8d1dd
(cherry picked from commit 4e99b444a00b3fb694040a2d1f1599d122394ced)

GitOrigin-RevId: e125b1d70fe52ebfff1460d6f5f14f3dbe84b388