Limit width of custom config dropdown
To avoid custom config name of extreme length
preventing the recording options from resizing,
a limit has been placed on the dropdown width.

This was done through setting the prototype display
value (the value used to initially calculate the
width of the dropdown). It was set to a minimum length
so it is always overriden/superceeded by the width of its
parent component, maintaining the past behavior and UI on
resizing, but limiting the length of the custom options.

Bug: 243638431
Test: Manual
Change-Id: I5efcb881d9d2b85114a8ad2b9ea5864ed8d2435c

GitOrigin-RevId: ca0b7d5a2bbd2c7598be15d4add669221a815539