Convert Notification to Kotlin
This updates Notification from Java to Kotlin. This allowed us to make
use of data classes which automatically generate the equals and hash
code functions for us.

As part of this change, the builder class was removed as it was not
added enough value in comparison to the amount of code it added.

Change-Id: Iaf3676d63da47778481657e6ed77f59c4cd9f22a

GitOrigin-RevId: c9f7c7b379a92c09f4f573d7f503d394af2b2204