Update model files now that RenderScript is disabled by default
This change also modifies IdeModelSnapshotComparisonTestDefinition to
stop comparing model v1 to model v2 for AGP 8.0 because model v1 is no
longer used in AGP 8.0. (Model v1 includes the RenderScript sources even
when renderscript is disabled, so the model v1 to model v2 comparison
was failing).

This change also modifies AndroidGradleTestUtils to alow tests to
specify whether to include RenderScript source files (false by default).
This functionality is needed for the tests in
RenderScriptDefaultRefactoringProcessorTest.

This change also changes the version in
RenderScriptDefaultRefactoringProcessor.necessity() from 8.0.0-alpha04
to 8.0.0-alpha02, because that will be the version of AGP in which
RenderScript will be disabled by default.

Bug: 219515633
Bug: 197900429
Test: existing
Change-Id: I557d5c82a24ccdb815f82b5913e7651bceefbfe6

GitOrigin-RevId: 0d61529ee53aabf3e64634d93690c897a05dc82a