Reflect the layout editor configuration to validation tool
When the configuration of layout editor is changed, reflects the
"effective change" to the validation tool.

The "effective change" means changing it doesn't break the intention
of the current configuration sets.

For example, when "Reference Devices" is selected:
- Changing the night mode is an effective change and it should reflect
  to the validation tool.
- Changing the device is NOT an effective change, because the
  intention of "Reference Devices" is to show the different devices
  of the current layout editor configuration. All the configuration
  attributes are inherit from layout editor except the device. Thus,
  changing device in layout editor should not reflect to validation
  tool.

Same, when "Project Locales" is selected:
- Changing the locale is NOT an effective change.
- Changing other configuration attributes are the effective changes.
  (e.g. orientation, device, target api level, night mode, etc)

This policy applies to other configuration sets.

Note that there is no effective change for the "Custom" set. Since
all the attributes of a custom configuration is defined by user, no
attribute is inherited from the layout editor.

Bug: 243628715
Test: manaully and added
Change-Id: I36bb3ac669e74847af731ac83300e6c8418bf466

GitOrigin-RevId: 5ef7b76a72afc4ac82e4c2b132d5968848ff0270