Add timeout to BLeak dominator phase
If there are many leak roots, computing the nodes dominated by the leak
root children can get expensive (~2.5s per leak root), which can cause
the entire test to time out. When this happens it produces no useful
information.

This change introduces a timeout on this phase (30s by default), so that
the maximum time it takes is predictable, and at least the basic
information for all the leak roots will be available.

Change-Id: I45da3f26088081cdf3de5d6d929e35fdc7e8afe3

GitOrigin-RevId: 6d0d89f376bf9d2361dbae073085a2742318614d