Fix occasional missing AndroidVersion
The existing logic to get the debug device's Android version was more
roundabout than it had to be; it used the overall XDebuggerManager to
get the current debug process and fetch the data from there. That debug
process could be null depending on timing, leading to a missing Android
version.

AndroidPositionManager already has its debug process passed into its
constructor, so by using it directly we ensure we can always get the
version correctly.

Test: Included
Bug: 74128300
Change-Id: Iae12e61b5163e58fb1d5ab8182fe282065f73823

GitOrigin-RevId: 240a6928e30d6ce8cde16d6a8ec3974b0719e4ab