Use module dir as initial destination
... for bundle/APK signing when signing a module for the first time.

When a user signs for the first time any module of a project, we store
the destination folder in the project properties, so if a user
customizes it then they do not have to select it again. But if a user
wants to sign another module, then we keep using the previous
destination, even if it is inside other module's root folder.

This change replaces a global property with a property for each module
and target type, so any customization for that particular combination
is kept.

Bug: 119886985
Test: ApkStetTest, GradleSignStepTest, manually checked path changed.
Change-Id: I4ae436e151dac8cebc4727bdcf8548c249883a85

GitOrigin-RevId: e23ab6b8376ff9ab1d78dbe68581882b0fe05867