Fix race condition that creates duplicate resources.
Currently it is possible for the StringResourceEditor to
create duplicate copies of resources. This occurs because
of rapid keystrokes and a ListenableFuture - the keystrokes
create duplicate futures, both of which create a new resource.

To remedy this, we simply keep track of whether we're currently
about to create a new resource, and if so, we ignore the second
event. The longer-term solution to this is to use bindings and
not key listeners.

Fixes: 216490744
Test: Synchronization issue probably not worth testing
Change-Id: Ib6a0f814060a1025ba004e59fc51972b667fe479

GitOrigin-RevId: 669042651ac5b4bc7188fcdc2cf683e29799d368