Refactor Compose-related build actions
Unify ForceCompileAndRefreshActionForNotification with its predecessor
and parent (ForceCompileAndRefreshAction) now that our UX supports only
one of them.

Also, make ComposePreviewRepresentation register a shortcut to
BuildAndRefresh as well, which is the action used in the preview context
when using fast preview. ForceCompileAndRefreshActionForNotification is
still used to display the action icon next to the status component when
fast preview is disabled, so the users can click on it to trigger a
build as an alternative to using the shortcut.

Fixes: 245825829
Test: compose-designer tests should pass
Change-Id: I85ebb2f8e1add0aba3eec41cbb39a4ba18df4f3c

GitOrigin-RevId: 114ca2a4c3cfd98b5142fa4a344dfb7acb07e6ab